Minimal audio implementation
I investigated various implementations of sound in other emulators,
which tended to be a bit tricky. After brainstorming with @ladyada
we hit upon the idea of simply assuming samples are written as
bytes in order and trapping on the write to the RAM address of the
final sample to collect a full 370 sample buffer.

This is good enough to play the system beep. It also works with some
games (dark castle) but not others (glider) and still does not
successfully play samples in hypercard. This may have to do with
the incomplete implementation of the VIA timer2, because debug
messages about timer2 are printed when trying to play sounds in
hypercard.