fix mounting discs when called for
by not clearing the accrun flag, the dialog to insert a previously
ejected disc successfully causes the emulated mac to cycle to the
correct disc.

This worked in several cases I tested, though in the specific case
of an older OS (About Finder says 5.3) with a newer floppy
(HyperCardBootSystem7.img) the hypercard floppy keeps adding a new
desktop icon everytime it's ejected then re-loaded, and the request
to load it can never be satisfied.