unix_main: Support configuring modifier keys
The new `-k` argument takes some comma-separated assignments, such
as `-k ctrl=option,alt=command`. In this configuration, the SDL
control keys are mapped to the mac option key, and the SDL alt keys
are mapped to the Mac command key. The SDL "gui" keys remain
associated with the Mac command key, and so the (non-existent on mac
plus) Mac control key cannot be pressed with this particular
configuration.

When `-k` is not specified, the behavior is intended to be unchanged.